# openapi.yaml
components:
  schemas:
    AvatarDeleteResponse:
      description: Schema for avatar deletion response
      properties:
        deleted:
          description: Whether avatar was successfully deleted
          title: Deleted
          type: boolean
        message:
          description: Status message
          title: Message
          type: string
        user_id:
          description: User ID
          format: uuid
          title: User Id
          type: string
      required:
      - user_id
      - deleted
      - message
      title: AvatarDeleteResponse
      type: object
    AvatarResponse:
      description: Schema for avatar response
      properties:
        avatar_url:
          description: Avatar URL
          title: Avatar Url
          type: string
        created_at:
          description: Creation timestamp
          format: date-time
          title: Created At
          type: string
        file_size:
          anyOf:
          - type: integer
          - type: 'null'
          description: File size in bytes
          title: File Size
        filename:
          anyOf:
          - type: string
          - type: 'null'
          description: Original filename
          title: Filename
        id:
          description: Unique identifier
          format: uuid
          title: Id
          type: string
        thumbnail_sizes:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          description: Available thumbnail sizes with their URLs
          title: Thumbnail Sizes
        updated_at:
          description: Last update timestamp
          format: date-time
          title: Updated At
          type: string
        upload_date:
          description: Upload timestamp
          format: date-time
          title: Upload Date
          type: string
        user_id:
          description: User ID
          format: uuid
          title: User Id
          type: string
      required:
      - created_at
      - updated_at
      - id
      - user_id
      - avatar_url
      - upload_date
      title: AvatarResponse
      type: object
...
    Body_upload_avatar_api_v1_users__user_id__avatar_post:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
      - file
      title: Body_upload_avatar_api_v1_users__user_id__avatar_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: "\n        # AI Ticket Creator Backend API\n        \n        Comprehensive\
    \ API for AI-powered ticket creation and management with Chrome extension support.\n\
    \        \n        ## Features\n        \n        - **AI-Powered Ticket Creation**:\
    \ Smart ticket creation with context analysis\n        - **File Processing**:\
    \ Upload and process multiple file types with AI transcription/OCR\n        -\
    \ **Real-time Notifications**: WebSocket-based real-time updates\n        - **Third-party\
    \ Integrations**: Salesforce, Jira, ServiceNow, Zendesk, GitHub, Slack, Teams,\
    \ Zoom\n        - **Chrome Extension Support**: Optimized for browser extension\
    \ workflows\n        - **Multi-Tenant Organization Support**: Complete organization\
    \ isolation for enterprise deployments\n        \n        ## Quick Start Guide\n\
    \        \n        ### 1. Authentication\n        Register a new user with organization:\n\
    \        ```bash\n        curl -X POST \"http://localhost:8000/api/v1/auth/register\"\
    \ \\\n          -H \"Content-Type: application/json\" \\\n          -d '{\"email\"\
    : \"user@company.com\", \"full_name\": \"John Doe\", \"password\": \"SecurePass123\"\
    , \"organization_name\": \"Acme Corporation\"}'\n        ```\n        \n     \
    \   ### 2. Add JIRA Integration\n        ```bash\n        curl -X POST \"http://localhost:8000/api/v1/integrations\"\
...
    \ and meeting management\n        "
...
openapi: 3.1.0
paths:
...
  /api/v1/users/{user_id}/avatar:
    delete:
      description: 'Delete a user''s avatar.


        - **user_id**: ID of the user to delete avatar for

        - Requires authentication

        - Users can only delete their own avatars unless they are admin'
      operationId: delete_avatar_api_v1_users__user_id__avatar_delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarDeleteResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Delete Avatar
      tags:
      - Users
    get:
      description: 'Get a user''s avatar image.


        - **user_id**: ID of the user

        - **size**: Avatar size (small, medium, large) - defaults to medium

        - Returns the actual image file

        - No authentication required for viewing avatars'
      operationId: get_avatar_api_v1_users__user_id__avatar_get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: query
        name: size
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: medium
          title: Size
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Avatar
      tags:
      - Users
    post:
      description: 'Upload a new avatar for a user.


        - **user_id**: ID of the user to upload avatar for

        - **file**: Image file to upload (JPG, PNG, GIF, HEIC, WebP)

        - Requires authentication

        - Users can only upload their own avatars unless they are admin'
      operationId: upload_avatar_api_v1_users__user_id__avatar_post
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_avatar_api_v1_users__user_id__avatar_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Upload Avatar
      tags:
      - Users
  /api/v1/users/{user_id}/avatar/info:
    get:
      description: 'Get avatar information (metadata only, not the image file).


        - **user_id**: ID of the user

        - Returns avatar metadata including URLs for different sizes

        - Requires authentication'
      operationId: get_avatar_info_api_v1_users__user_id__avatar_info_get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: Get Avatar Info
      tags:
      - Users
...